Add action Panel
