--- ui/app/model/ActionExtensionButton.java
@@ -1,25 +1,0 @@
-package model;
-
-public class ActionExtensionButton {
-
-	public ActionExtensionButton(){
-	}
-
-	private String _displayName;
-	public String getDisplayName(){
-		return _displayName;
-	}
-	public void setDisplayName(String value){
-		_displayName = value;
-	}
-
-	private String _fontAwesomeName;
-	public String getFontAwesomeName(){
-		return _fontAwesomeName;
-	}
-
-	public void setFontAwesomeName(String value){
-		_fontAwesomeName = value;
-	}
-
-}
