+++ ui/app/views/node/detail.scala.html
@@ -21,7 +21,7 @@
 			</span>
 		</h4>
 	</div>
-	
+
 	<ul class="nav nav-tabs">
   		<li @if(!activatePreviewTab){class="active"}><a href="@routes.Node.showProperty(repositoryId, obj.getId)" data-toggle="tab" data-target="#property">@Messages("view.node.property")</a></li>
   		@if(Util.isDocument(obj)){
@@ -33,14 +33,20 @@
 	  		<li><a href="@routes.Node.showVersion(repositoryId, obj.getId)" data-toggle="tab" data-target="#version">@Messages("view.node.version")</a></li>
   		}
   		<li><a href="@routes.Node.showPermission(repositoryId, obj.getId)" data-toggle="tab" data-target="#permission">@Messages("view.node.permission")</a></li>
+	  	@for(actionPluginUI <- Util.getActionPluginUIElementList(obj)){
+	  		<li><a href="@routes.Node.showAction(repositoryId, obj.getId, actionPluginUI.getActionId )" data-toggle="tab" data-target="#action-@actionPluginUI.getActionId()">@actionPluginUI.getActionId()</a></li>
+	  	}
 	</ul>
- 
+
 	<div class="tab-content">
 	  	<div class="tab-pane fade @if(!activatePreviewTab){active in}" id="property"></div>
 	  	<div class="tab-pane fade" id="file"></div>
 	  	<div class="tab-pane fade" id="version"></div>
 	  	<div class="tab-pane fade @if(activatePreviewTab){active in}" id="preview"></div>
 	  	<div class="tab-pane fade" id="permission"></div>
+	  	@for(actionPluginUI <- Util.getActionPluginUIElementList(obj)){
+	  		  	<div class="tab-pane fade" id="action-@actionPluginUI.getActionId()"></div>
+	  	}
 	</div>
 
 
@@ -70,8 +76,8 @@
 		     }
 		 });
 	}
-	
-	 
+
+
 	 $('a[data-toggle="tab"]').on('shown.bs.tab', function (e) {
 		  var target = $(e.target).attr("href"); // activated tab
 		    $.ajax({
@@ -86,7 +92,7 @@
 		      }
 		  });
 		})
-		
+
 		$('.favorite-toggle').on('click', function(){
 			$.ajax({
 				type: "PUT",
@@ -102,8 +108,8 @@
 					$('#favorite-star').attr('title', "@Messages("view.user.favorite.toggled.desc1")" + "\n" + "@Messages("view.user.favorite.toggled.desc2")");
 				}
 			}).fail(function(xhr, status, error){
-				alert(xhr.status + " " + xhr.statusText + ": Failed to swith favorite star");	
+				alert(xhr.status + " " + xhr.statusText + ": Failed to swith favorite star");
 			});
 		});
 	</script>
-} 
+}
